+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v0.2.0
+
+- Remove necessity to load the worker script separately
+- Remove `workerscriptLocation` option
+- Added support for Gnomad SV data
+- Added `maxTileWidth` to control who far the user can zoom out.
 
 v0.1.0